Commit: Criação do bd e adicionando as tabelas

+++ src/main/java/br/com/arthouseserv/arthouseserv/dto/ContatoDTO.java
@@ -2,5 +2,5 @@ package br.com.arthouseserv.arthouseserv.dto;
 
 import java.time.LocalDateTime;
 
-public record ContatoDTO(Long id, String email, String celular, LocalDateTime dataHoraCriacao, LocalDateTime dataHoraAlteracao) {
+public record ContatoDTO(Integer id, String email, String celular, LocalDateTime dataHoraCriacao, LocalDateTime dataHoraAlteracao) {
 }
