+++ src/main/java/br/com/arthouseserv/arthouseserv/dto/LeadDTO.java
@@ -3,5 +3,5 @@ package br.com.arthouseserv.arthouseserv.dto;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public record LeadDTO(Long id, String nome, LocalDateTime dataHoraCriacao, LocalDateTime dataHoraAlteracao, List<ContatoDTO> contatos) {
+public record LeadDTO(Integer id, String nome, LocalDateTime dataHoraCriacao, LocalDateTime dataHoraAlteracao, List<ContatoDTO> contatos) {
 }
